beautify quitting message
When running as a client, and the application ends but not caused by
user input (C-c), the there is no carriage return after the exit
message, and so the 'quitting' message will drag the terminal prompt
forward.

This patch fixes that cosmetic bug by adding a carriage return to the
quitting message.